Add xml payload to link
Modifies the web_server used in poc_two to serve a page containing a
link that has a payload for the gateway defined in the data attribute.

This payload would be generated on the server and rendered to the page,
likely as a signed JWT that we can verify in the gateway (as per
poc_one). JWT has not been implemented here for simplicity.

The payload is JSON with the format:

{
  "type":"xml",
  "payload":"<?xml version=\"1.0\" encoding=\"UTF-8\"?><methodCall> <methodName>DisplayStudy</methodName>
              <params> <param> <value> <struct> <member> <name>accessionNumber</name>
              <value> <string>123456</string> </value> </member> </struct> </value>
              </param> </params> </methodCall>",
  "destination":"http://localhost:8081"
}

There is a likelihood that gateway will need to pass various message
formats e.g. hl7, FHIR, JSON, xml etc to various systems within the
site. If we use something like this payload format we can hopefully keep
the gateway as simple as possible and configure any customisations
within the web application.